move parallel gateway tests to their own package
