+++ tests/SpiffWorkflow/bpmn/parallel_gateway_tests/BaseParallelTestCase.py
