+++ tests/SpiffWorkflow/bpmn/parallel_gateway_tests/ParallelGatewayTest.py
@@ -2,7 +2,7 @@
 
 from SpiffWorkflow.task import TaskState
 from SpiffWorkflow.bpmn.workflow import BpmnWorkflow
-from .BpmnWorkflowTestCase import BpmnWorkflowTestCase
+from ..BpmnWorkflowTestCase import BpmnWorkflowTestCase
 
 __author__ = 'matth'
