+++ tests/SpiffWorkflow/bpmn/parallel_gateway_tests/ParallelLoopingAfterJoinTest.py
@@ -1,8 +1,7 @@
 # -*- coding: utf-8 -*-
 
-import unittest
 from SpiffWorkflow.bpmn.workflow import BpmnWorkflow
-from tests.SpiffWorkflow.bpmn.BaseParallelTestCase import BaseParallelTestCase
+from .BaseParallelTestCase import BaseParallelTestCase
 
 __author__ = 'matth'
 
@@ -23,8 +22,3 @@ class ParallelLoopingAfterJoinTest(BaseParallelTestCase):
             ['Go', '1', '2', '2A', '2B', '2 Done', ('Retry?', 'Yes'), 'Go',
                      '1', '2', '2A', '2B', '2 Done', ('Retry?', 'No'), 'Done'], save_restore=True)
 
-
-def suite():
-    return unittest.TestLoader().loadTestsFromTestCase(ParallelLoopingAfterJoinTest)
-if __name__ == '__main__':
-    unittest.TextTestRunner(verbosity=2).run(suite())
